safari-600.7-branch/LayoutTests/http/tests/css/css-image-valued-shape-expected.txt

Verify that a image valued shape-outside layout is updated after the image has been loaded. This test checks that left edge of the "Hello World" text span is defined by the 200px wide float before shape-outside image has been loaded and by the 100px wide image after it's been laoded.

Hello World
PASS elementRect("text").left is 200
PASS elementRect("text").left is 100
PASS successfullyParsed is true

TEST COMPLETE
